requirements.txt
Flask==1.1.1
pandas==0.24.2
google-cloud-storage==1.19.1
gunicorn==19.9.0
sklearn==0.0